buildfarm_wizard_config/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
catkin-pkg==0.4.16
click==6.7
distro==1.4.0
docutils==0.16
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
gitdb2==2.0.5
GitPython==2.1.11
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.3.0
pathlib2==2.3.2
pluggy==0.7.1
py==1.6.0
pyparsing==2.4.6
pytest==3.8.0
python-dateutil==2.8.1
PyYAML==5.3
rosdistro==0.8.0
rospkg==1.2.3
ruamel.yaml==0.15.66
six==1.11.0
smmap2==2.0.5
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
